Re-add `addIdentifiers`
Per feedback on the pull request, I've re-added `addIdentifiers()`. This
also required deferring listener preparation to the first `trigger()`,
which resolves the "order of operations" issues that led to removal of
`addIdentifiers` in the first place.

Additionally, I've refactored and simplified the logic for listener
preparation. First, I've added an `attachListenerStructs()` method which
loops through a list of listener structs and attaches them to the given
event; this allowed removal of 3 code duplications. Next, I updated
`attach()` to push wildcard events into a special array; this allows
deferring wildcard attachment until all listeners are registered, and
also simplifies attaching wildcard listeners from the shared manager
(they are aggregated at the end).

The changes pose no change in the performance improvements seen with the
original refactor!