Implemented lazy listeners
This patch creates a "lazy listener" feature. Usage is:

```php
// Individual listener:
$events->attach('foo', (new LazyListener($definition, $container))->getListener());

// Aggregates:
$events->attachAggregate(new LazyListenerAggregate($listeners, $container));
```

`$definition` should be a struct defining the following:

- event (event name to which to attach; ignored when attaching a lazy listener
  directly to the event manager)
- listener (listener service name by which to pull the listener from the
  container)
- method (listener method to invoke for the event)
- priority (optional; priority at which to attach the instance; ignored when
  attaching a lazy listener directly to the event manager)

For aggregates, `$listeners` can be an array of `LazyListener` instances, or
array structs defining parameters for a `LazyListener` instance (in the latter
case, the aggregate will create the instance for you).

Internally, `LazyListener::getListener()` creates a closure around retrieval of
the listener and retrieval of the listener method; it then invokes the method on
the listener, passing the event, and returning the value.