Distinguish between lazy listeners for aggregates
Lazy listeners are useful for direct attachment:

```php
$events->attach('foo', new LazyListener($definition, $container));
```

and, as such, do not need to know the event or priority (as those are
explicitly referenced in the `attach()` invocation).

When attaching via an aggregate, however, we want to know the event and
optionally the priority. As such, we need to distinguish between them.

This patch removes the event and priority awareness from `LazyListener`,
and creates a new class, `LazyEventListener`, which extends
`LazyListener` and adds event and priority awareness.
`LazyListenerAggregate` now aggregates instances of the new class.