Remove isPrepared logic
Removes the isPrepared logic. zend-mvc in v2 makes extensive use of the idea
that shared listeners can be added at any time in the lifecycle; if an instance
cannot identify those changes, then all shared listeners must effectively be
added prior to the first trigger of `Zend\Mvc\Application`.

This patch removes the logic, and adds logic that does the following:

- When pulling shared listeners, it does it based on the current event being
  triggered.
- It checks to see if it has pulled shared listeners previously:
  - If not, it caches the structs and returns them.
  - If the previously registered structs are identical, it does nothing.
  - If they are different, it determines if any were removed, and detaches them;
    any new ones are then returned so they may be attached. The current set is
    then cached.

The only behavior change is that we now allow adding/setting identifiers after
first preparation, so the logic to guard against that and related tests were
removed.

Performance is almost identical to the previous commit (e1e1441)!