Implement TaskProcessorInterface via EventManager
Updates the `EventManagerInterface` to also extend
`TaskProcessorInterface` and `ListenerProviderInterface`.

`getListenersForEvent()` takes over most of the logic from
`triggerListeners()` with regard to aggregating and sorting the
listeners, and is implemented as a generator. For listener
discovery/matching, it uses the results of `$event->getName()` if the
event is a package-specific event, and the name is set; otherwise, it
uses the class name as the event name.

`triggerListeners()` now iterates over `getListenersForEvent()` to
trigger listeners and aggregate results. This change is transparent to
end users.

`process()` iterates over `getListenersForEvent()` as well, but does not
aggregate results, and instead returns the event.

Essentially, these become two different triggering paradigms.