feat: Creates PrioritizedListenerProvider
New provider implements `PrioritizedListenerAttachmentInterface` and
`PrioritizedListenerProviderInterface`, and will iterate attached
listeners in priority order.

Each iteration will take into account both the event name, if a
`getName()` method is available, the event class, and any wildcard
listeners, and listeners of the same priority will be returned in the
order they are attached, based on those criteria.